backend/requirements.txt
cement>=3.0.10
Jinja2==3.1.2
PyYAML==6.0.1
colorlog==6.8.0
xlrd==2.0.1
mysql-connector-python==8.3.0
APScheduler==3.10.4

akshare>=1.18.0
tushare==1.3.6
lxml==5.1.0
pandas==2.1.4
requests==2.28.1
bs4==0.0.1

Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
django-extensions>=3.2
whitenoise>=6.6
gunicorn>=21.0
